requirements.txt
python-dotenv
streamlit
langchain-google-genai
python-dotenv
SpeechRecognition
pyaudio
